# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Centros",
        "description": "Documentação da API de Centros, incluindo operações de CRUD e relacionamentos.",
        "version": "1.1.0"
    },
    "servers": [
        {
            "url": "/api",
            "description": "API principal"
        }
    ],
    "paths": {
...
        "/centros": {
            "get": {
                "tags": [
                    "Centros"
                ],
                "summary": "Listar todos os centros",
                "operationId": "b51ec8c659378cd0561a10d6a6ac0522",
                "responses": {
                    "200": {
                        "description": "Lista de centros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Centro"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Centros"
                ],
                "summary": "Criar um novo centro",
                "operationId": "a58e977e93fc6924791d82cb4d2fc781",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CentroInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Centro criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Centro"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/centros/{id}": {
            "get": {
                "tags": [
                    "Centros"
                ],
                "summary": "Buscar centro por ID",
                "operationId": "9d949b35eb79adc43e59923eb70f5686",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Centro encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Centro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Centro não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Centros"
                ],
                "summary": "Atualizar centro",
                "operationId": "ef17826017a578d79d9bfb338be5935e",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CentroInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Centro atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Centro"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Centro não encontrado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Centros"
                ],
                "summary": "Deletar centro",
                "operationId": "415235a399d7ee1d84f3c17d343064e1",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Centro deletado"
                    },
                    "404": {
                        "description": "Centro não encontrado"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
            "Centro": {
                "required": [
                    "nome",
                    "localizacao",
                    "contactos"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "readOnly": true
                    },
                    "nome": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "localizacao": {
                        "type": "string"
                    },
                    "contactos": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    }
                },
                "type": "object"
            },
            "CentroInput": {
                "required": [
                    "nome",
                    "localizacao",
                    "contactos"
                ],
                "properties": {
                    "nome": {
                        "type": "string",
                        "maxLength": 100
                    },
                    "localizacao": {
                        "type": "string"
                    },
                    "contactos": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "email": {
                        "type": "string",
                        "format": "email",
                        "nullable": true
                    }
                },
                "type": "object",
                "example": {
                    "nome": "Centro de Formação Profissional",
                    "localizacao": "Rua das Flores, 123, Luanda",
                    "contactos": [
                        "923456789",
                        "912345678"
                    ],
                    "email": "contato@centro.com"
                }
            },
...
        }
    },
...
}